--- src/main/java/com/kodilla/ecommercee/GenericEntityRepository.java
@@ -1,7 +1,0 @@
-package com.kodilla.ecommercee;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface GenericEntityRepository extends JpaRepository<GenericEntity, Long> { }
