+++ src/main/java/com/kodilla/ecommercee/domain/GenericEntity.java
@@ -1,0 +1,31 @@
+package com.kodilla.ecommercee.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class GenericEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String value;
+
+    public GenericEntity() {
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public Long getId() {
+
+        return id;
+    }
+
+    public GenericEntity(String value) {
+
+        this.value = value;
+    }
+}
