--- src/main/java/com/kodilla/ecommercee/dto/ProductDto.java
@@ -1,27 +1,0 @@
-package com.kodilla.ecommercee.dto;
-
-public class ProductDto {
-    private Long id;
-    private String name;
-
-    public ProductDto(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
