--- src/main/java/com/kodilla/ecommercee/dto/UserDto.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.dto;
-
-public class UserDto {
-}
