+++ src/main/java/com/kodilla/ecommercee/repository/GenericEntityRepository.java
@@ -1,0 +1,8 @@
+package com.kodilla.ecommercee.repository;
+
+import com.kodilla.ecommercee.domain.GenericEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GenericEntityRepository extends JpaRepository<GenericEntity, Long> { }
